Changes requested for ip check
