+++ elements/package.json
@@ -10,6 +10,7 @@
   "license": "MIT",
   "dependencies": {
     "@ucd-lib/theme-sass": "^5.0.15",
+    "ip-cidr": "^3.0.4",
     "lit": "^2.0.2",
     "photoswipe": "^4.1.3",
     "slim-select": "^1.26.2"
