+++ elements/utils/controllers/index.js
@@ -4,7 +4,6 @@ import { MutationObserverController } from "./mutation-observer";
 import { PopStateObserverController } from "./popstate-observer";
 import { SilsPrimoController } from "./sils-primo";
 import { WaitController } from "./wait";
-import { IpDetect} from "./ip-check";
 
 
 export {
@@ -14,5 +13,4 @@ export {
   PopStateObserverController,
   SilsPrimoController,
   WaitController,
-  IpDetect,
 };
