+++ test-app/package.json
@@ -32,6 +32,7 @@
     "dotenv": "^16.0.0",
     "express": "^4.17.1",
     "fs-extra": "^10.0.0",
+    "ip-cidr": "^3.0.4",
     "lit": "^2.0.2",
     "marked": "^2.1.3",
     "photoswipe": "^4.1.3",
